+++ tools/playground/index.html
@@ -6,16 +6,12 @@
     <link rel="icon" href="data:,">
 
     <script src="libs/ace.js" type="text/javascript" charset="utf-8"></script>
-    <script src="../owl.js"></script>
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/solid.css" integrity="sha384-QokYePQSOwpBDuhlHOsX0ymF6R/vLk/UQVz3WHa6wygxI5oGTmDTv8wahFOSspdm" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/fontawesome.css" integrity="sha384-vd1e11sR28tEK9YANUtpIOdjGW14pS87bUBuOIoBILVWLFnS+MCX9T6MMf0VdPGq" crossorigin="anonymous">
-
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/solid.css" integrity="sha384-QokYePQSOwpBDuhlHOsX0ymF6R/vLk/UQVz3WHa6wygxI5oGTmDTv8wahFOSspdm" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/fontawesome.css" integrity="sha384-vd1e11sR28tEK9YANUtpIOdjGW14pS87bUBuOIoBILVWLFnS+MCX9T6MMf0VdPGq" crossorigin="anonymous">
+    
     <!-- Application JS/CSS -->
     <link rel="stylesheet" href="playground.css">
-    <script type="module" src="export_snippet.js"></script>
-    <script type="module" src="utils.js"></script>
-    <script type="module" src="app.js"></script>
-
+    <script type="module" src="playground.js"></script>
   </head>
   <body>
   </body>
