+++ tools/playground/samples/custom_hooks/custom_hooks.css
@@ -1,0 +1,6 @@
+
+button {
+    width: 120px;
+    height: 35px;
+    font-size: 16px;
+}
