+++ tools/playground/samples/jsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+        "paths": {
+            "owl": ["../../../src/index.ts"],
+        },
+    }
+}
