+++ tools/playground/samples/lifecycle_demo/lifecycle_demo.css
@@ -1,0 +1,11 @@
+button {
+    font-size: 18px;
+    margin: 5px;
+}
+
+.demo {
+    margin: 10px;
+    padding: 10px;
+    background-color: #dddddd;
+    width: 250px;
+}
